refactor: query_checker 대신 rag score를 사용하도록 변경
