+++ src/agent/agent.py
@@ -4,7 +4,6 @@ from langgraph.graph.state import CompiledStateGraph
 from langgraph.prebuilt import tools_condition
 
 from agent.nodes.cc_agent import cc_agent
-from agent.nodes.query_checker import query_checker
 from agent.nodes.query_retriever import query_retriever
 from agent.nodes.tool_node import tool_node
 from agent.utils.state import AgentState
@@ -13,13 +12,14 @@ from agent.utils.state import AgentState
 def get_graph() -> CompiledStateGraph:
     builder = StateGraph(AgentState)
     builder.add_node("query_retriever", query_retriever)
-    builder.add_node("query_checker", query_checker)
+    # builder.add_node("query_checker", query_checker)
     builder.add_node("cc_agent", cc_agent)
     builder.add_node("tools", tool_node)
 
     builder.add_edge(START, "query_retriever")
-    builder.add_edge("query_retriever", "query_checker")
-    builder.add_edge("query_checker", "cc_agent")
+    # builder.add_edge("query_retriever", "query_checker")
+    # builder.add_edge("query_checker", "cc_agent")
+    builder.add_edge("query_retriever", "cc_agent")
     builder.add_conditional_edges("cc_agent", tools_condition, ["tools", END])
     builder.add_edge("tools", "cc_agent")
