+++ src/agent/tools/company_info_tool.py
@@ -6,8 +6,8 @@ from langchain.tools import tool
 @tool
 def company_info_tool() -> Dict:
     """
-    회사의 정보를 제공합니다.
-    사용자가 필요한 문서나 정보를 얻을 수 있도록 이메일과 담당자 전화번호를 제공합니다.
+    회사의 이메일과 담당자 전화번호를 제공합니다.
+    사용자가 제출해야하는 문서, 서류가 있을 경우에 이 정보를 활용할 수 있습니다.
 
     Returns:
         Dict: {
