+++ src/agent/tools/express_carrier_tool.py
@@ -1,44 +1,8 @@
-from enum import Enum
 from typing import Dict
 
 from langchain.tools import tool
 
-
-class ExpressCourier(Enum):
-    PATECH = "파테크"
-    SGL = "SGL"
-    NAMGYEONG = "남경"
-    LOGISTORM = "로지스톰"
-    LOTOS = "LOTOS"
-    ACE = "ACE"
-
-
-CARRIER_INFO = {
-    ExpressCourier.PATECH: {
-        "email": "admin@forwarder.kr",
-        "Telephone": "032-201-1155",
-    },
-    ExpressCourier.SGL: {
-        "email": "sgl@siriusglobal.co.kr",
-        "Telephone": "051-441-7341",
-    },
-    ExpressCourier.NAMGYEONG: {
-        "email": "gtjeon@namkyungglobal.com",
-        "Telephone": "02-577-3331",
-    },
-    ExpressCourier.LOGISTORM: {
-        "email": "logistormail@gmail.com",
-        "Telephone": "02-2667-0306",
-    },
-    ExpressCourier.LOTOS: {
-        "email": "sale@lotos.co.jp",
-        "Telephone": "+81-3-6278-9408",
-    },
-    ExpressCourier.ACE: {
-        "email": "import2@iecoz.com",
-        "Telephone": "02-2038-7224",
-    },
-}
+from agent.utils.constants import CARRIER_INFO, ExpressCourier
 
 
 @tool
